Move the condition payload into conditiongql
Every other shared GraphQL payload in the provider lives in its own
package (aiagentgql, fieldconditiongql, fieldgql, pipegql and the rest),
so the condition payload and its selection belong in conditiongql rather
than bundled with the Terraform-side model. conditionschema keeps what
the provider owns: the tfsdk model, the schema attributes, and the
mapping in both directions. fieldconditiongql embeds
conditiongql.Condition, and the automation reads through the same
selection.

Also dropped two comments that restated the code they sat above.
